[CP] Add Redux dependencies to redux-provider
We add all npm packages related to Redux that need to be used by
the projects inside this example.

These packages will be exported through Liferay AMD Loader and
made available to the rest of projects using the bundler's imports
feature.

When you deploy this checkpoint commit (see README) open your browser's
console and execute the following Javascript:

  Liferay.Loader.require("redux-provider$redux@4.0.1", console.log)

You should get the dump of an object (Redux's export) in the console.
That shows that this provider is making Redux available through the AMD
loader.