[CP] Create empty Redux store
This store will be shared among all portlets contained in this example
so that they can communicate with each other, maintaining the global
state of the UI in a centralized singleton JavaScript module.

There are other techniques to communicate portlets like IPC, the window
object, Liferay events, etc., but given that Redux fits very well in
React's application model we choose this one.

We will then add actions and reducers to the store as needed.

When you deploy this checkpoint commit (see README) open your browser's
console and execute the following Javascript:

  Liferay.Loader.require("redux-store@1.0.0", console.log)

You should get the dump of the recently created Redux store which should
be an object with a `default` field pointing to an object with the usual
`dispatch()` and `subscribe()` methods, typical of Redux stores.