Create input-data-widget project
Run yo liferay-script to create a React Widget project.

This project will contain a widget that renders an UI to gather data
that is then rendered by a different widget.

The data will be in CSV format and will consist of body measures
(weight, fat, muscle, ...) and it will be transferred from the input to
the display widget by means of the Redux store.