Import external dependencies in input-data-widget
We will import React, Redux and the common Redux store from the other
projects in this example.

Note that we remove React from package.json as we don't need it for
anything during build time.

If we were writing tests or using  Typescript, for example, we would
perhaps need to leave those dependencies in package.json for the build
tools to find them.