[CP] Create an empty AppComponent widget
This will be the initial bootstrap component that will be attached to
the DOM.

For now it just renders an empty content and makes sure that the Redux
store is provided to all its descendants.

When you deploy this checkpoint commit (see README) a widget should
appear in Liferay's `Add` > `Widgets` menu, inside a category named
`Liferay JS Toolkit Showcase`.

Go ahead and add an instance of that widget to the page. It will render
empty, but it will be updated in following checkpoint commits.