Define configuration for input-data-widget
Because our input widget will show a form to input CSV data, we will
create a configuration option that can be set per portlet instance to
fill the form with initial data. That way the user won't need to type
anything to make things work.

Note that the configuration labels are specified as symbols. We will
provide multi language values for those symbols in the following
commits.