Add a formData field to Redux store
Because our input widget is going to show an input area, we need to
back up its data in a store field. Thus, we set an initial value for the
field and create an action plus a reducer to modify it.

Note how we have created a merge function to avoid modifying the
Redux store directly and, instead, make a copy of it to be returned by
the reducer. This is a requisite of Redux and is usually achieved with
different npm packages, but to keep this example simple, we will do
it with a simple local function.