Load initial data from configuration
Now that we have a Redux field to store input data, we will get the
initial value from the portlet instance configuration and dump it into
the Redux store.

Note that because the configuration field is a simple input text and we
need multiline data, we replace '|'  by CRs as a way to let users
specify where CRs occur.