[CP] Render input form in input-data-widget
After you deploy this checkpoint commit (see README) the `Input Data
Widget` widget will show a text area where you can enter CSV data with
body measures.

And because you entered some data in the configuration dialog in a
previous checkpoint commit, it will be automatically filled.

Now, remember how we added a `formData` field to the Redux store in a
previous commit? Open your browser's console and run the following code:

  Liferay.Loader.require(
    "redux-store@1.0.0",
    store => console.log(store.default.getState())
  )

This will dump the contents of the Redux store because it is provided by
the `redux-store@1.0.0` module as the default export and it contains a
`getState()` method that lets you inspect the contents of the store as
it is now.

In this case you should see the dump of an object with two fields named
`data` and `formData`, where the `formData` one has the CSV data you
entered in the `Input Data Widget`.