Create display-data-widget project
Run yo liferay-script to create a React Widget project.

This project will contain a widget that renders an UI to display data
that is provided by another widget.

The data will be in CSV format and will consist of body measures
(weight, fat, muscle, ...) and it will be transferred from the input
widget to this widget by means of the Redux store.