Define a system wide configuration for display-data-widget
Because our display widget will show a graph to display data, we will
create a configuration option that can be set system wide to define the
initial display options for the graph.

Note that the configuration labels are specified as symbols. We will
provide multi language values for those symbols in the following
commits.