[CP] Add Spanish L10N to the project
Note that, in addition to the labels file, we register es_ES in
.npmbuildrc. This is not mandatory, but if you want to use the translate
task in the future, it will make it work without warning you.

When you deploy this checkpoint commit (see README) the `Display Data
Widget` widget should appear in Liferay's `Add` > `Widgets` menu, inside
a category named `Liferay JS Toolkit Showcase`.

Go ahead and add an instance of that widget to the page. It will render
empty, but it will be updated in following checkpoint commits.

Now, because you have defined a system wide configuration for that
widget, you can go to `Control Panel` > `Configuration` > `System
Settings`, scroll down to the `Other` category and click on the wheel
icon of this widget's configuration to define its initial values.

When you do that, a configuration dialog with `Default measure` and
`Period` fields will appear. Go ahead and configure some values there,
then click `Update`. Those values will be used as defaults when the
widget if fully rendered in the following commits.