[CP] Create React based project
Run yo liferay-js to create a React based portlet and add a button with
a red border to the sample code created by the Yeoman generator.

Note that the CSS code to put the red border around the button comes
from the styles.css file in the assets folder and is loaded by the
portlet itself when instantiated.