[CP] Add a Material-UI demo
Demonstrate how to use Material-UI icons and styles engine.

Note that we need to disable the `replace-browser-modules` plugin for
the `jss-*` packages because their `browser` field points to an invalid
.js file (in the sense that it tries to use the keyword `export` instead
of `module.exports` as expected by the bundler).

Also note that we import icons explicitly by file path. There's an
alternative syntax in Material-UI to import icons directly from
`@material-ui/icons` but because we use an AMD loader, it causes one
module definition per icon which is quite poor in terms of performance.